requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
deep-translator==1.11.4
python-dotenv==1.0.0
google-generativeai==0.8.5
supabase>=2.0.0
